ProductFinderWebApp/requirements.txt
Flask==1.1.2
gunicorn==20.1.0
eventlet==0.30.2
gevent==21.1.2
numpy==1.20.2
Pillow==8.2.0
pypng==0.0.20